+++ nintendeals/noe/api/__init__.py
@@ -1,4 +1,6 @@
 from .nintendo import search_by_nsuid
+from .nintendo import search_dlcs_by_platform
 from .nintendo import search_by_platform
 from .nintendo import search_by_query
+from .nintendo import search_recent_switch_dlcs
 from .nintendo import search_recent_switch_games